requirements.txt
streamlit>=1.20.0
langgraph>=0.0.37
google-generativeai>=0.4.0
transformers>=4.30.0
torch>=1.13.0
requests>=2.28.0
pandas>=1.5.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.2
scikit-learn
fastapi>=0.95.0
uvicorn>=0.22.0
